Tidied up test suite types and naming
