+++ src/ui/components/TestSuite/utils/isTestSuiteReplay.ts
@@ -1,10 +1,13 @@
 import { Recording } from "shared/graphql/types";
-import { isGroupedTestCases, isIncrementalGroupedTestCases } from "shared/test-suites/types";
+import {
+  isGroupedTestCasesV2,
+  isGroupedTestCasesV3,
+} from "shared/test-suites/RecordingTestMetadata";
 
 export function isTestSuiteReplay(recording: Recording): boolean {
   const testMetadata = recording?.metadata?.test;
   return (
     testMetadata != null &&
-    (isIncrementalGroupedTestCases(testMetadata) || isGroupedTestCases(testMetadata))
+    (isGroupedTestCasesV2(testMetadata) || isGroupedTestCasesV3(testMetadata))
   );
 }
